Otimização do React Navigation
Problemas de conflitos de navegadores foi resolvido. Todas as páginas bem definidas para evitar futuros problemas de navegação pelas rotas.
Foram implementadas muitas novidades de arquivos e códigos como preparação para conexões e autenticações futuras.
