+++ talent-ui/package.json
@@ -26,6 +26,7 @@
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-svg": "13.4.0",
+    "realm": "^11.10.1",
     "typescript": "^4.9.4"
   },
   "devDependencies": {
