+++ talent-ui/src/databases/schemas/LoginSchema.ts
@@ -1,0 +1,10 @@
+export const LoginSchema = {
+  name: "Login",
+  properties: {
+    _id: "string",
+    name: "string",
+    last_name: "string",
+    email: "string",
+    password: "string"
+  }
+}
