+++ talent-ui/src/services/index.tsx
@@ -1,0 +1,14 @@
+// import React, { useContext } from 'react'
+
+// import AuthContext from '../contexts/auth'
+
+// import AppRoutes from './auth.signup'
+// import AuthRoutes from './auth.login'
+
+// const Routes: React.FC = () => {
+//   const {signed} = useContext(AuthContext);
+
+//   return signed ? <AppRoutes /> : <AuthRoutes />;
+// };
+
+// export default Routes;
